requirements.txt
flask==2.3.2
gunicorn==21.2.0
google-cloud-aiplatform==1.48.0
python-dotenv==1.0.0
pandas==2.2.3
numpy==1.26.4